Basics of what we're after.
Returns list of column count length.

Each item in list is a dictionary of non-unique values mapped to
offending rows.

There's still some question of stripping col values per row.

Also still need be able to include/exclude columns on the command line and
as a parameter to the search function.